fix(aod): minor improvements
