+++ .vscode/settings.json
@@ -1,1 +1,5 @@
-{}
+{
+  "cSpell.words": [
+    "Populator"
+  ]
+}
