--- libs/application/templates/announcement-of-death/src/fields/RealEstateAndLandsRepeater/index.tsx
@@ -1,1 +1,0 @@
-export { RealEstateAndLandsRepeater } from './RealEstateAndLandsRepeater'
