+++ libs/application/templates/announcement-of-death/src/forms/done.ts
@@ -2,6 +2,7 @@ import {
   buildForm,
   buildCustomField,
   buildMultiField,
+  buildDescriptionField,
 } from '@island.is/application/core'
 import { Form, FormModes } from '@island.is/application/types'
 import { m } from '../lib/messages'
@@ -14,9 +15,14 @@ export const done: Form = buildForm({
     buildMultiField({
       id: 'done',
       title: m.announcementComplete,
-      description: m.nextStepsText,
+      description: m.announcementCompleteDescription,
       space: 1,
       children: [
+        buildDescriptionField({
+          id: 'nextSteps',
+          title: '',
+          description: m.nextStepsText,
+        }),
         buildCustomField({
           id: 'completeStepImage',
           title: '',
