+++ libs/application/templates/announcement-of-death/src/forms/draft/subSectionInfo.ts
@@ -3,6 +3,7 @@ import {
   buildSelectField,
   buildTextField,
   buildSubSection,
+  buildDescriptionField,
 } from '@island.is/application/core'
 import { m } from '../../lib/messages'
 import { Application } from '../../types/schema'
