+++ libs/application/templates/announcement-of-death/src/lib/utils/index.ts
@@ -27,6 +27,7 @@ export const getFileRecipientName = (
   const estateMember = estateMembers?.members.find(
     (estateMember) => estateMember.nationalId === recipient,
   )
+
   return estateMember?.name || answers.applicantName.toString()
 }
